EE-871: move pure PoS logic into separate crate
