+++ execution-engine/contracts/system/pos/src/contract_runtime.rs
@@ -1,19 +1,6 @@
 use contract::contract_api::runtime;
+use proof_of_stake::RuntimeProvider;
 use types::{account::PublicKey, key::Key, BlockTime, Phase};
-
-pub trait RuntimeProvider {
-    fn get_key(name: &str) -> Option<Key>;
-
-    fn put_key(name: &str, key: Key) -> ();
-
-    fn remove_key(name: &str) -> ();
-
-    fn get_phase() -> Phase;
-
-    fn get_block_time() -> BlockTime;
-
-    fn get_caller() -> PublicKey;
-}
 
 #[allow(dead_code)]
 pub struct ContractRuntime;
