+++ execution-engine/proof-of-stake/src/stakes_provider.rs
@@ -1,0 +1,7 @@
+use crate::{stakes::Stakes, Result};
+
+pub trait StakesProvider {
+    fn read() -> Result<Stakes>;
+
+    fn write(stakes: &Stakes);
+}
